crated optypes.py (#207)
