+++ coreblocks/params/__init__.py
@@ -1,4 +1,5 @@
-from .genparams import *  # noqa: F401
 from .isa import *  # noqa: F401
+from .optypes import *  # noqa: F401
+from .genparams import *  # noqa: F401
 from .layouts import *  # noqa: F401
 from .fu_params import *  # noqa: F401
