+++ test/frontend/test_decoder.py
@@ -1,9 +1,8 @@
-from amaranth import *
 from amaranth.sim import *
 
 from ..common import TestCaseWithSimulator, test_gen_params
 
-from coreblocks.params.isa import *
+from coreblocks.params import *
 from coreblocks.frontend.decoder import InstrDecoder
