+++ test/fu/test_jb_unit.py
@@ -4,8 +4,6 @@ from parameterized import parameterized_class
 
 from coreblocks.params import *
 from coreblocks.fu.jumpbranch import JumpBranchFuncUnit, JumpBranchFn, JumpComponent
-from coreblocks.params.fu_params import FunctionalComponentInputs, FunctionalComponentParams
-from coreblocks.params.isa import OpType
 from coreblocks.transactions.lib import Method, def_method
 from coreblocks.params.layouts import FuncUnitLayouts, FetchLayouts
 from coreblocks.utils.protocols import FuncUnit
